requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
appdirs=1.4.4=pypi_0
asgiref=3.7.2=pypi_0
asttokens=2.4.1=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.2=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py311h6a678d5_7
build=1.0.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.22=pypi_0
click=8.1.7=pypi_0
colorama=0.4.4=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=4.7.2=pypi_0
cryptography=41.0.7=py311hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-runtime=12.1.0=0
dataclasses-json=0.6.3=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.14=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=7.0.0=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.109.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
flaml=2.1.1=pypi_0
flatbuffers=23.5.26=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
fuzzywuzzy=0.18.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
googleapis-common-protos=1.62.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httptools=0.6.1=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=py311h06a4308_0
importlib-metadata=6.11.0=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipython=8.21.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=py311h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
kubernetes=29.0.0=pypi_0
lame=3.100=h7b6447c_0
langchain=0.1.4=pypi_0
langchain-community=0.0.16=pypi_0
langchain-core=0.1.16=pypi_0
langsmith=0.0.83=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
levenshtein=0.24.0=pypi_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.10.1=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.2=pypi_0
markupsafe=2.1.3=py311h5eee18b_0
marshmallow=3.20.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.26.3=py311h08b1b3b_0
numpy-base=1.26.3=py311hf175353_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.17.0=pypi_0
openai=0.28.1=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
opentelemetry-api=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.22.0=pypi_0
opentelemetry-instrumentation=0.43b0=pypi_0
opentelemetry-instrumentation-asgi=0.43b0=pypi_0
opentelemetry-instrumentation-fastapi=0.43b0=pypi_0
opentelemetry-proto=1.22.0=pypi_0
opentelemetry-sdk=1.22.0=pypi_0
opentelemetry-semantic-conventions=0.43b0=pypi_0
opentelemetry-util-http=0.43b0=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.0.1=py311ha6cbd5a_0
pip=23.3.1=py311h06a4308_0
plotly=5.18.0=pypi_0
posthog=3.4.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pulsar-client=3.4.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyautogen=0.1.14=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pygments=2.17.2=pypi_0
pyopenssl=23.2.0=py311h06a4308_0
pypdf=4.0.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.0.0=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
python-levenshtein=0.24.0=pypi_0
pytorch=2.1.2=py3.11_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-lightning=2.1.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=py311h5eee18b_0
rapidfuzz=3.6.1=pypi_0
readline=8.2=h5eee18b_0
recbole=1.2.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=py311h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.39.2=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py311h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.36.3=pypi_0
sympy=1.12=py311h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
texttable=1.7.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.2=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=pypi_0
torchaudio=2.1.2=py311_cu121
torchmetrics=1.3.0.post0=pypi_0
torchtriton=2.1.0=py311
torchvision=0.16.2=py311_cu121
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
transformers=4.37.1=pypi_0
typer=0.9.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2023.4=pypi_0
urllib3=1.26.18=py311h06a4308_0
uvicorn=0.27.0.post1=pypi_0
uvloop=0.19.0=pypi_0
wandb=0.16.2=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.7.0=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
